여러가지 package화 및 facebook 가입이후 로그인시 적용이 안되는 부분을 수정
