+++ .requirements/dev.txt
@@ -27,6 +27,7 @@ psycopg2-binary==2.7.4
 ptyprocess==0.5.2
 Pygments==2.2.0
 python-dateutil==2.6.1
+python-magic==0.4.15
 pytz==2018.3
 PyYAML==3.12
 raven==6.6.0
