+++ app/members/apis/__init__.py
@@ -1,106 +1,6 @@
-from django.contrib.auth import get_user_model
-from rest_framework import generics, status, permissions
-from rest_framework.authtoken.models import Token
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from .permissions import IsOwnerOrReadOnly
-from .serializers import AccessTokenSerializer, UserSerializer, EmailAuthTokenSerializer
+from .list import UserListView
+from .detail import UserDetailView
+from .singup import SignupView
+from .login import AuthTokenForFacebookAccessTokenView, AuthTokenForEmailView
+from .logout import LogoutView
 
-User = get_user_model()
-
-__all__ = (
-    'AuthTokenForFacebookAccessTokenView',
-    'UserList',
-    'UserDetail',
-    'Logout',
-    'AuthTokenForEmailView',
-    'SignupView',
-)
-
-
-class AuthTokenForEmailView(APIView):
-    def post(self, request):
-        serializer = EmailAuthTokenSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        user = serializer.validated_data['user']
-        token, is_create = Token.objects.get_or_create(user=user)
-        data = {
-            'token': token.key,
-            'user': UserSerializer(user).data,
-        }
-        return Response(data)
-
-
-class AuthTokenForFacebookAccessTokenView(APIView):
-    def post(self, request):
-        serializer = AccessTokenSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        user = serializer.validated_data['user']
-        token, _ = Token.objects.get_or_create(user=user)
-        data = {
-            'token': token.key,
-            'user': UserSerializer(user).data,
-        }
-        return Response(data)
-
-
-class Logout(APIView):
-    permission_classes = (permissions.IsAuthenticated,)
-
-    def get(self, request, format=None):
-        request.user.token.delete()
-        return Response(status=status.HTTP_200_OK)
-
-
-class SignupView(APIView):
-    def post(self, request):
-        serializer = UserSerializer(data=request.data)
-        if serializer.is_valid():
-            User.objects.create_user(
-                email=serializer.initial_data['email'],
-                nickname=serializer.initial_data['nickname'],
-                password=serializer.initial_data['password'],
-            )
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class UserList(generics.ListAPIView):
-    queryset = User.objects.filter(is_active=True, is_staff=False, is_superuser=False)
-    serializer_class = UserSerializer
-
-
-class UserDetail(generics.RetrieveUpdateDestroyAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-
-    permission_classes = (
-        permissions.IsAuthenticatedOrReadOnly,
-        IsOwnerOrReadOnly,
-    )
-
-
-
-
-# class UserList(APIView):
-#     def get(self, request, format=None):
-#         users = User.objects.all()
-#         serializer = UserSerializer(users, many=True)
-#         return Response(serializer.data)
-#
-#     def post(self, request, format=None):
-#         serializer = UserSerializer(data=request.data)
-#         if serializer.is_valid():
-#             serializer.save()
-#             return Response(serializer.data, status=status.HTTP_201_CREATED)
-#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-#
-#
-# class UserDetail(APIView):
-#     permission_classes = (permissions.IsAuthenticated,)
-#
-#     def get(self, request):
-#         # user = self.request.user
-#         serializer = UserSerializer(request.user)
-#         return Response(serializer.data)
