+++ app/members/backends.py
@@ -1,8 +1,10 @@
 import requests
 from django.contrib.auth import get_user_model
+from django.core.files import File
 from rest_framework import status
 
 from config import settings
+from utils.file import download, get_buffer_ext
 
 User = get_user_model()
 
@@ -28,14 +30,12 @@ class APIFacebookBackend:
             response_dict = response.json()
             facebook_id = response_dict['id']
             url_picture = response_dict['picture']['data']['url']
-            user, _ = User.objects.get_or_create(username=facebook_id, img_profile=url_picture)
-
+            user, _ = User.objects.get_or_create(username=facebook_id)
+            if not user.img_profile:
+                temp_file = download(url_picture)
+                ext = get_buffer_ext(temp_file)
+                user.img_profile.save(f'{user.pk}.{ext}', File(temp_file))
             return user
-            # if not user.img_profile:
-            #     temp_file = download(url_picture)
-            #     ext = get_buffer_ext(temp_file)
-            #     user.img_profile.save(f'{user.pk}.{ext}', File(temp_file))
-            # return user
 
 
     def get_user(self, user_id):
