+++ app/members/serializers/email_serializer.py
@@ -1,0 +1,40 @@
+from django.utils.translation import ugettext_lazy as _
+from rest_framework import serializers
+from django.contrib.auth import get_user_model
+from rest_framework.compat import authenticate as rest_authenticate
+
+User = get_user_model()
+
+__all__ = (
+    'EmailAuthTokenSerializer',
+)
+
+
+class EmailAuthTokenSerializer(serializers.Serializer):
+    email = serializers.CharField(label=_("Email"))
+    password = serializers.CharField(
+        label=_("Password"),
+        style={'input_type': 'password'},
+        trim_whitespace=False
+    )
+
+    def validate(self, attrs):
+        email = attrs.get('email')
+        password = attrs.get('password')
+
+        if email and password:
+            user = rest_authenticate(request=self.context.get('request'),
+                                     email=email, password=password)
+
+            # The authenticate call simply returns None for is_active=False
+            # users. (Assuming the default ModelBackend authentication
+            # backend.)
+            if not user:
+                msg = _('Unable to log in with provided credentials.')
+                raise serializers.ValidationError(msg, code='authorization')
+        else:
+            msg = _('Must include "email" and "password".')
+            raise serializers.ValidationError(msg, code='authorization')
+
+        attrs['user'] = user
+        return attrs
